Move notifications to gov_users
