+++ conf/urls.py
@@ -37,7 +37,6 @@ urlpatterns = [
     path("queries/", include("queries.urls")),
     path("routing-rules/", include("workflow.routing_rules.urls")),
     path("licences/", include("licences.urls")),
-    path("menu/", include("menu.urls")),
     re_path(r"^swagger(?P<format>\.json|\.yaml)$", schema_view.without_ui(cache_timeout=0), name="schema-json",),
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
